refactoring to support premium time-period logic
The manager should be able to see all the periods the user was premium.
This is done via a list of tuples (a, b), where a is always
the start date and b the end date of a premium period.
This allows the calculation of the bill for any given month, even if premium was activated and terminated multiple times in one month (hopefully).